% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_blaster.R
\name{comma_blaster}
\alias{comma_blaster}
\title{Comma-Separated Strings for Vector Definition}
\usage{
comma_blaster(x, sep = " ", char_out = FALSE, spaces = TRUE)
}
\arguments{
\item{x}{a character string to be comma-separated.}

\item{sep}{element in the input string to be replaced with commas.
Defaults to " ".}

\item{char_out}{logical; should the output be printed with quotes around each 
vector element?}

\item{spaces}{logical; should each comma be followed by a space (just for looks)?}
}
\value{
a comma-separated character string that can be copied from the
  console and used to define a vector.
}
\description{
Format R console printouts for definition of new vectors.
  Sometimes faster than figuring out how to extract vectors from complex
  objects. Often useful during exploratory phase of analysis.
}
\examples{
> comma_blaster(month.abb[1:5], char_out=TRUE)
[1] "'Jan', 'Feb', 'Mar', 'Apr', 'May'"
> newvec <- c('Jan', 'Feb', 'Mar', 'Apr', 'May')
> 
> comma_blaster("1-800-555-0100", sep='-')
[1] "1, 800, 555, 0100"
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\seealso{
\code{\link{excel_to_r}}
}
\keyword{'data}
\keyword{creation'}

