% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_r.R
\name{excel_to_r}
\alias{excel_to_r}
\title{Comma-Separated Strings from Spreadsheet Cell Vectors}
\usage{
excel_to_r(char_out = FALSE, spaces = TRUE, ...)
}
\arguments{
\item{char_out}{logical; Should the output be printed with quotes 
around each vector element?}

\item{spaces}{logical; Should each comma be followed by a space 
(just for looks)?}

\item{...}{additional arguments to \code{scan}.}
}
\value{
a comma-separated string that can be copied from the console 
  and used to define a vector.
}
\description{
Paste values from spreadsheets into R in vector-ready
  (comma separated) format.
}
\examples{
excel_to_r(spaces=FALSE)

#not run:
Paste Excel values into console, then hit ENTER
1: 0    3   3.5 2.5 3   4   2.5 0
9: 
Read 8 items
[1] "0,3,3.5,2.5,3,4,2.5,0"

#run:
newvec <- c(0,3,3.5,2.5,3,4,2.5,0)


excel_to_r(char_out=TRUE, allowEscapes=FALSE)

#not run:
Paste Excel values into console, then hit ENTER
1: 4    6   3.2 \\n
5: 
Read 4 items
[1] "'4', '6', '3.2', '\\\\n'"
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\seealso{
\code{\link{comma_blaster}}
}
\keyword{'data}
\keyword{creation'}

