% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matfilter.R
\name{matfilter}
\alias{matfilter}
\title{Proportion-Based Data Subsetting}
\usage{
matfilter(x, mar = 2, cond = NA, fun = NULL, thresh = 1, na.rm = TRUE,
  filter = TRUE)
}
\arguments{
\item{x}{the matrix or data.frame to be subset.}

\item{mar}{the margin along which the function will be applied - either
1 for rows or 2 for columns.}

\item{cond}{character expression; the completion of a Boolean expression,
e.g. ">= 0", which is passed to \code{apply}.}
}
\value{
a matrix or data.frame with columns or rows removed according to
  the condition or function specified.
}
\description{
Remove rows or columns according to the proportion of elements that 
  meet the condition specified.
}
\examples{
> df <- data.frame(a=c(2,2,3,3), b=c(1,2,NA,1), c=c(2,NA,NA,3))
> df
  a  b  c
1 2  1  2
2 2  2 NA
3 3 NA NA
4 3  1  3

> matfilter(df, mar=2, fun=is.na, thresh=0.3)
$column_proportions
   a    b    c 
0.00 0.25 0.50 

  a  b
1 2  1
2 2  2
3 3 NA
4 3  1

> matfilter(df, mar=1, cond='== 2', thresh=0.5)
$row_proportions
[1] 0.6666667 0.6666667 0.0000000 0.0000000

  a  b  c
3 3 NA NA
4 3  1  3 
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{slicing}
\keyword{subsetting,}

