% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcol_inserter.R
\name{rowcol_inserter}
\alias{rowcol_inserter}
\title{Insert a Row or Column Into a Matrix or Data Frame}
\usage{
rowcol_inserter(x, vec, pos, name = pos, dim = "row")
}
\arguments{
\item{x}{data.frame or matrix to modify.}

\item{vec}{a vector to be inserted.}

\item{pos}{the numerical position in the data.frame 
or matrix where the vector will be inserted.}

\item{name}{a character name for the row or column (optional).}

\item{dim}{'row' to insert a row, 'col' to insert a column. 
Defaults to 'row'.}
}
\value{
a data.frame or matrix with row or column inserted.
}
\description{
Insert a Row or Column Into a Matrix or Data Frame
}
\examples{
#example dataframe
x <- data.frame('col1'=1:10, 'col2'=rnorm(10), 'col3'=letters[1:10],
stringsAsFactors=FALSE)

#insert new row at position 8
rowcol_inserter(x, c(99,100,'new'), 8)

#insert new column at position 3
rowcol_inserter(x, 101:110, 3, name='newcol', dim='col')
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{'data}
\keyword{manipulation'}

