% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultimate_reader.R
\name{ultimate_reader}
\alias{ultimate_reader}
\title{Read and Merge Multiple Files}
\usage{
ultimate_reader(dir_args = list(path = "./", pattern = ".csv"),
  read_args = list(sep = ",", quote = "\\"", header = TRUE, fill = TRUE,
  comment.char = ""), merge = FALSE, ...)
}
\arguments{
\item{dir_args}{\code{list} of args passed to \code{dir}, which selects
the files to be read/merged.}

\item{read_args}{\code{list} of args passed to \code{read.table}, 
which by default is set up to read .csv files.}

\item{merge}{determines whether to merge each file and output the merged
\code{data.frame} (TRUE) or leave them separate and add them to the
global environment (FALSE).}

\item{...}{supplies additional args to \code{merge}.}
}
\value{
either returns a merged \code{data.frame} or adds individual
  \code{data.frame}s to the global environment.
}
\description{
Reads any character-delimited filetype based on filename patterns.
  Merges data if desired.
}
\examples{
ultimate_reader(dir_args=list(path='~/stuff/', pattern='\\w{3}_2016.csv')

merged <- ultimate_reader(merge=TRUE, by='date', all=TRUE)
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{merge}
\keyword{read}

